requirements.txt
Python 3.8+
PyTorch
OpenCV
NumPy
PIL
Streamlit
torchvision